+++ lib/khiva.rb
@@ -4,6 +4,7 @@ require "fiddle/import"
 # modules
 require "khiva/array"
 require "khiva/clustering"
+require "khiva/dimensionality"
 require "khiva/distances"
 require "khiva/features"
 require "khiva/library"